Thêm mã giảm giá

+++ lab03/Repositories/EFProductRepository.cs
@@ -15,8 +15,17 @@ namespace lab03.Repositories
             // return await _context.Products.ToListAsync();
             return await _context.Products
             .Include(p => p.Category) // Include thông tin về category
+            .Include(p => p.Variants)
             .ToListAsync();
         }
+        public async Task<List<Product>> GetAllWithVariantsAsync()
+        {
+            return await _context.Products
+                .Include(p => p.Category)
+                .Include(p => p.Variants)
+                .ToListAsync();
+        }
+
         public async Task<Product> GetByIdAsync(int id)
         {
             // return await _context.Products.FindAsync(id);
@@ -35,7 +44,28 @@ namespace lab03.Repositories
         }
         public async Task DeleteAsync(int id)
         {
-            var product = await _context.Products.FindAsync(id);
+            var product = await _context.Products
+        .Include(p => p.Images)
+        .FirstOrDefaultAsync(p => p.Id == id);
+
+            if (product == null)
+            {
+                throw new Exception("Product not found");
+            }
+
+            // Kiểm tra nếu sản phẩm đã từng xuất hiện trong OrderDetail
+            bool hasOrderDetails = await _context.OrderDetails.AnyAsync(od => od.ProductId == id);
+            if (hasOrderDetails)
+            {
+                throw new Exception("Không thể xóa sản phẩm vì đã có đơn hàng liên quan.");
+            }
+
+            // Xóa ảnh liên quan (nếu có)
+            if (product.Images != null && product.Images.Any())
+            {
+                _context.ProductImages.RemoveRange(product.Images);
+            }
+
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
         }
